ppdb/test/black/requirements.txt
python-memcached>=1.59 